Schematic for BMS companion board
This is the schematic for the BMS to Inverter PCB. It's yet to communicate with either the BMS or inverter but works with an Arduino Uno with an RS485 shield or a Can bus shield so there should be no reason why this will not work. It will be fully tested when a replacement BMS is received and an inverter purchased. The circuit board otherwise seems to work perfectly.